Add files via upload
We aim to demonstrate the use of create, read, update and delete operations through this project. The project starts by adding a cab and details of its driver .The owner provides cab to the drivers and ads their expenses on daily basis. Booking scene is where customers can book a taxi to get to the desired location. On booking, the passenger receives information like booking date & time , pick up & drop locations, and details of the cab & the driver to the registered mobile number and email address. At the end of the ride a bill will generated with the total distance traveled, time taken and total amount to be pain. The passenger can choose from any of the available payment methods .